requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pandas>=2.1.0
numpy>=1.26.0
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
yfinance>=0.2.40
